requirements.txt
aiocontextvars==0.2.2
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
contextvars==2.4
Django==3.1.3
django-celery-beat==2.1.0
django-timezone-field==4.0
djangorestframework==3.12.2
idna==2.10
immutables==0.14
importlib-metadata==2.0.0
kombu==4.6.11
loguru==0.5.3
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.0
